requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==23.2.1
    # via gradio
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   gradio
    #   httpx
    #   openai
    #   starlette
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   streamlit
    #   typer
    #   uvicorn
distro==1.9.0
    # via openai
fastapi==0.115.6
    # via gradio
ffmpy==0.4.0
    # via gradio
filelock==3.16.1
    # via huggingface-hub
fsspec==2024.10.0
    # via
    #   gradio-client
    #   huggingface-hub
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
gradio==5.9.0
    # via llm-sport (pyproject.toml)
gradio-client==1.5.2
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   openai
    #   safehttpx
huggingface-hub==0.27.0
    # via
    #   gradio
    #   gradio-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   gradio
    #   pydeck
jiter==0.8.2
    # via openai
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.18.4
    # via altair
numpy==2.2.0
    # via
    #   gradio
    #   opencv-python
    #   pandas
    #   pydeck
    #   streamlit
openai==1.57.4
    # via llm-sport (pyproject.toml)
opencv-python==4.10.0.84
    # via llm-sport (pyproject.toml)
orjson==3.10.12
    # via gradio
packaging==24.2
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   streamlit
pandas==2.2.3
    # via
    #   gradio
    #   streamlit
pillow==11.0.0
    # via
    #   gradio
    #   streamlit
protobuf==5.29.1
    # via streamlit
pyarrow==18.1.0
    # via streamlit
pydantic==2.10.3
    # via
    #   fastapi
    #   gradio
    #   openai
pydantic-core==2.27.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pydub==0.25.1
    # via gradio
pygame==2.6.1
    # via llm-sport (pyproject.toml)
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via llm-sport (pyproject.toml)
python-multipart==0.0.19
    # via gradio
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   llm-sport (pyproject.toml)
    #   huggingface-hub
    #   streamlit
rich==13.9.4
    # via
    #   streamlit
    #   typer
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
ruff==0.8.3
    # via gradio
safehttpx==0.1.6
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.41.3
    # via
    #   fastapi
    #   gradio
streamlit==1.41.1
    # via llm-sport (pyproject.toml)
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tomlkit==0.13.2
    # via gradio
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer==0.15.1
    # via gradio
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typer
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
uvicorn==0.34.0
    # via gradio
websockets==14.1
    # via gradio-client
